requirements.txt
aioredis==1.2.0
appnope==0.1.0
asgiref==3.1.4
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-environ==0.4.5
django-extensions==2.1.9
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.2
gunicorn==19.9.0
haikunator==2.1.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.18
incremental==17.5.0
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
Markdown==3.1.1
more-itertools==7.1.0
msgpack==0.6.1
nose==1.3.7
packaging==19.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
pprint==0.1
prettyprint==0.1.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pyparsing==2.4.0
pytest==5.0.1
pytest-asyncio==0.10.0
pytest-django==3.5.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.2
zipp==0.5.2
zope.interface==4.6.0
